feat: refactor staff management by moving Staff struct to states module
